% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_data.R
\name{anova_feature}
\alias{anova_feature}
\title{Perform an ANOVA test for a single feature}
\usage{
anova_feature(df, df.groups, feature_name, plot_path, prefix = "")
}
\description{
Perform an ANOVA test for a single feature
}
