% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_data.R
\name{hmCreateCellDataset}
\alias{hmCreateCellDataset}
\title{Creates a `CellDataSet` object from a `heteromotility` object}
\usage{
hmCreateCellDataset(hm, scalar = "data", cat = c("clust"))
}
\arguments{
\item{hm}{: `hetermotility` data object.}

\item{scalar}{: character. scalar data set to use. {'data', 'unscaled.data', 'pcs'}.}

\item{cat}{vector of characters. properties from `@meta.data` to include in `phenoData`.}
}
\value{
cds : `CellDataSet` object
}
\description{
Creates a `CellDataSet` object from a `heteromotility` object
}
