% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_data.R
\name{hmGetData}
\alias{hmGetData}
\title{Gets data from a `heteromotility` object and returns a data.frame}
\usage{
hmGetData(hm, scalar = "data", cat = c("clust"))
}
\arguments{
\item{hm}{: `heteromtility` data object.}

\item{scalar}{: string. property for numerical data to be retrieved. {'data', 'raw.data', 'pcs', etc}}

\item{cat}{: iterable of strings. categorical features to retrieve from `meta.data`}
}
\value{
df : data.frame. contains data in `scalar` and retrieved `meta.data`
}
\description{
Gets data from a `heteromotility` object and returns a data.frame
}
