% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_data.R
\name{hmHClust}
\alias{hmHClust}
\title{Performs hierarchical clustering and sets `meta.data$clust` in `hm`}
\usage{
hmHClust(hm, k = 2, linkage = "ward.D2", dist_method = "euclidean",
  scalar = "data")
}
\arguments{
\item{hm}{: `heteromtility` data object.}

\item{k}{: integer. number of clusters to set with `cutree`.}

\item{linkage}{: character. method for hierarchical clustering, compatible with `hclust`.}

\item{dist_method}{: character. method for distance matrix calculation, compatible with `dist`.}

\item{scalar}{: character. scalar data space to use. {'data', 'unscaled.data', 'pcs'}.}
}
\value{
hm : `heteromtility` data object.
  adds 'clust' variable to `hm@meta.data`
}
\description{
Performs hierarchical clustering and sets `meta.data$clust` in `hm`
}
