% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_data.R
\name{hmMakeObject}
\alias{hmMakeObject}
\title{Build a `heteromotility` data object from a data.frame}
\usage{
hmMakeObject(raw.data, meta.data.list = list())
}
\arguments{
\item{raw.data}{: data.frame of raw numeric data with 2 id columns}

\item{meta.data.list}{: list of character vectors or factors to include in meta.data}
}
\value{
hm : heteromotility data object.
}
\description{
Build a `heteromotility` data object from a data.frame
}
