% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_data.R
\name{hmPCA}
\alias{hmPCA}
\title{Performs PCA on `data` in a heteromotility object `hm`
adds PCs to `pcs` and loadings to `pc_loadings`}
\usage{
hmPCA(hm, max_pcs = 30)
}
\arguments{
\item{hm}{: `heteromtility` data object.}

\item{max_pcs}{: integer. number of principle components to keep.}
}
\value{
hm : `heteromtility` data object.
  assigns @pcs
}
\description{
Performs PCA on `data` in a heteromotility object `hm`
adds PCs to `pcs` and loadings to `pc_loadings`
}
