% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_plot.R
\name{hmPlotFeatureMeans}
\alias{hmPlotFeatureMeans}
\title{Plots feature means for each group defined by a categorical variable `class` in `meta.data`}
\usage{
hmPlotFeatureMeans(hm, class = "clust", limited = F, sig_bars = F,
  width = 8, height = 5, save = NULL)
}
\arguments{
\item{hm}{: `heteromtility` data object.}

\item{class}{: character. class in `meta.data` to group bars by.}

\item{limited}{: boolean. use a limited feature set for visualization.}

\item{sig_bars}{: boolean. plot significance bars. requires `class`_adj_p_value in `@feat.data`.}

\item{width}{: float. plot width for saving in inches.}

\item{height}{: float. plot width for saving in inches.}

\item{save}{: character. filename for saving.}
}
\value{
pl : ggplot2 object.
}
\description{
Plots feature means for each group defined by a categorical variable `class` in `meta.data`
}
