% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_plot.R
\name{hmPlotPCA}
\alias{hmPlotPCA}
\title{Plots PCA}
\usage{
hmPlotPCA(hm, pc1 = 1, pc2 = 2, color = "clust", save = NA)
}
\arguments{
\item{hm}{: `heteromtility` data object.}

\item{pc1}{: integer. first PC to plot.}

\item{pc2}{: integer. second PC to plot}

\item{color}{: string. name of a variable in meta.data to color points.}

\item{save}{: string. filename to save plots.}
}
\value{
pl : ggplot2 object.
}
\description{
Plots PCA
}
