% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_plot.R
\name{hmPlotPseudotime}
\alias{hmPlotPseudotime}
\title{Plots monocle pseudotiming}
\usage{
hmPlotPseudotime(hm, class = "State", show_state_number = F,
  rev_x_axis = F, save = NULL, width = 4, height = 4)
}
\arguments{
\item{hm}{: `heteromotility` object containing @celldataset trajectory data.}

\item{class}{: character. feature in FeatureData to color by.}

\item{show_state_number}{: boolean. show state numbers on pseudotiming plot.}

\item{rev_x_axis}{: boolean. reverse the order of the pseudotiming x-axis.}

\item{save}{: character. path to file for saving.}

\item{width}{: integer. Width of plots in inches.}

\item{height}{: integer. Height of plots in inches.}
}
\value{
pl : ggplot2 object.
}
\description{
Plots monocle pseudotiming
}
