% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_plot.R
\name{hmPlotTSNE}
\alias{hmPlotTSNE}
\title{Plots TSNE}
\usage{
hmPlotTSNE(hm, color = "clust", save = NA)
}
\arguments{
\item{hm}{: `heteromtility` data object.}

\item{color}{: string. name of a variable in meta.data to color points.}

\item{save}{: string. filename to save plots.}
}
\value{
pl : ggplot2 object.
}
\description{
Plots TSNE
}
