% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_data.R
\name{hmPseudotime}
\alias{hmPseudotime}
\title{Performs pseudotiming on motility feature data using `monocle`}
\usage{
hmPseudotime(hm, scalar = "data", cat = c("clust"))
}
\arguments{
\item{hm}{: `hetermotility` data object.}

\item{scalar}{: character. scalar data set to use. {'data', 'unscaled.data', 'pcs'}.}

\item{cat}{vector of characters. properties from `@meta.data` to include in `phenoData`.}
}
\value{
hm : `heteromtility` data object.
  assigns @celldataset.
}
\description{
Performs pseudotiming on motility feature data using `monocle`
}
