% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_data.R
\name{hmSubsetData}
\alias{hmSubsetData}
\title{Retrives a subset of cell data in heteromotility object `hm`}
\usage{
hmSubsetData(hm, cell_index = F, cat = F, id = F)
}
\arguments{
\item{hm}{: `heteromtility` data object.}

\item{cell_index}{: boolean or numerical index of rows in @data to keep.}

\item{cat}{: character. category to use when subsetting with `id`.}

\item{id}{: character. group in `cat` to keep.}
}
\value{
hm : `heteromtility` data object.
}
\description{
Useful for selecting only cells of a certain type or class.
`cat` and `id` parameters must be supplied together.
}
