% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_data.R
\name{hmTSNE}
\alias{hmTSNE}
\title{Performs tSNE on `pcs` in `hm`}
\usage{
hmTSNE(hm, perplexity = 50)
}
\arguments{
\item{hm}{: `heteromtility` data object.}

\item{perplexity}{: integer. perplexity parameter for `Rtsne`.}
}
\value{
hm : `heteromtility` data object.
  assigns `hm@tsne`
}
\description{
Performs tSNE on `pcs` in `hm`
}
