% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_data.R
\name{ttest_feature}
\alias{ttest_feature}
\title{Perform a t-test for a single feature}
\usage{
ttest_feature(df, df.class, feature_name, class1 = "Aged", class2 = "Young",
  plot_path = FALSE, prefix = "")
}
\description{
Perform a t-test for a single feature
}
