% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_data.R
\name{hmScaleData}
\alias{hmScaleData}
\title{Scales and centers `raw.data` in a heteromotility object `hm`}
\usage{
hmScaleData(hm)
}
\arguments{
\item{hm}{: `heteromtility` data object.}
}
\value{
hm : `heteromtility` data object.
  assigns @data.
}
\description{
Scales and centers `raw.data` in a heteromotility object `hm`
}
